[refactoring] 전략패턴 도입하여 코드를 좀 정리하였습니다
