+++ TestConsole/src/main/java/Main.java
@@ -1,17 +1,12 @@
-import script.ScriptRunner;
+import command.common.RunMode;
+import command.common.RunModeFactory;
+import command.console.TestConsole;
+import command.script.ScriptRunner;
 
 public class Main {
 
     public static void main(String[] args) {
-
-        if (args.length == 0) {
-            TestConsole console = new TestConsole();
-            console.run();
-        }
-
-        if(args.length == 1){
-            ScriptRunner scriptRunner = new ScriptRunner();
-            scriptRunner.run(args[0]);
-        }
+        RunMode mode = RunModeFactory.create(args);
+        mode.run();
     }
 }
