--- TestConsole/src/main/java/command/ExitCommand.java
@@ -1,22 +1,0 @@
-package command;
-
-public class ExitCommand extends Command {
-    public ExitCommand(ConsoleService service) {
-        super(service);
-    }
-    public static final int EXPECTED_ARGUMENT_COUNT = 1;
-
-    @Override
-    public String isValidArguments(String[] args) {
-        if (!isValidArgumentCount(args, EXPECTED_ARGUMENT_COUNT)) {
-            return INVALID_ARGUMENT_NUMBER_MSG;
-        }
-
-        return "";
-    }
-
-    @Override
-    public CommandResult doExecute(String[] args) {
-        return CommandResult.EXIT;
-    }
-}
