+++ TestConsole/src/main/java/command/common/CommandResult.java
@@ -1,4 +1,4 @@
-package command;
+package command.common;
 
 import logger.Logger;
