+++ TestConsole/src/main/java/command/common/ICommand.java
@@ -1,4 +1,4 @@
-package command;
+package command.common;
 
 public interface ICommand {
     CommandResult execute(String[] args);
