+++ TestConsole/src/main/java/command/common/RunMode.java
@@ -1,0 +1,5 @@
+package command.common;
+
+public interface RunMode {
+    void run();
+}
