+++ TestConsole/src/main/java/command/console/EraseCommand.java
@@ -1,4 +1,8 @@
-package command;
+package command.console;
+
+import command.common.Command;
+import command.common.CommandResult;
+import command.common.ConsoleService;
 
 public class EraseCommand extends Command {
     public static final int EXPECTED_ARGUMENT_COUNT = 3;
