+++ TestConsole/src/main/java/command/console/HelpCommand.java
@@ -1,4 +1,8 @@
-package command;
+package command.console;
+
+import command.common.Command;
+import command.common.CommandResult;
+import command.common.ConsoleService;
 
 public class HelpCommand extends Command {
     public HelpCommand(ConsoleService service) {
