+++ TestConsole/src/main/java/command/console/ReadCommand.java
@@ -1,4 +1,8 @@
-package command;
+package command.console;
+
+import command.common.Command;
+import command.common.CommandResult;
+import command.common.ConsoleService;
 
 public class ReadCommand extends Command {
     public ReadCommand(ConsoleService service) {
