+++ TestConsole/src/test/java/CommandPatternTest.java
@@ -1,4 +1,5 @@
-import command.*;
+import command.common.ConsoleService;
+import command.console.*;
 import org.junit.jupiter.api.Test;
 
 import static org.mockito.Mockito.*;
