+++ TestConsole/src/test/java/ConsoleServiceTest.java
@@ -1,6 +1,5 @@
-import command.ConsoleService;
+import command.common.ConsoleService;
 import jdk.jfr.Description;
-import org.junit.jupiter.api.Disabled;
 import org.junit.jupiter.api.RepeatedTest;
 import org.junit.jupiter.api.Test;
 import org.mockito.ArgumentCaptor;
