+++ TestConsole/src/test/java/DefaultTest.java
@@ -1,3 +1,4 @@
+import command.console.TestConsole;
 import org.junit.jupiter.api.Test;
 
 import java.io.*;
